ARTEMIS-5650 AMQP Federation receivers prefer modified dispositions by default
AMQP Federation receivers configure themselves to default to receive a modified
disposition for address full errors to allow the remote to redeliver the message
instead of the broker defaault which is to send rejected dispositions meaning the
remote must discard or DLQ the message. Allows for configuration at the federation
policy level to override this and ignore the connector URI as this is an opinionated
configuration for federation resources.